ARM: Rename CONFIG_SKIP_LOWLEVEL_INIT to CONFIG_SKIP_CPU_INIT_CRIT
For ARM, the CONFIG_SKIP_LOWLEVEL_INIT option implies that the
cpu_init_crit and lowlevel_init assembly functions are skipped. We may
want to skip only lowlevel_init, so rename the option that skips both
to CONFIG_SKIP_CPU_INIT_CRIT. The MIPS option of the same name
is not renamed.

Signed-off-by: Michael Spang <mspang@csclub.uwaterloo.ca>